Replace ll modifier in format strings with definable LL_FMT_STR constant, to be platform independent. Microsoft requires using I64 instead of ll.  Now we can specify it at build time depending on our build environment.

+++ 7/main.cpp
@@ -4,7 +4,7 @@
 
 #include <stdio.h>
 
-long main () {
+int main () {
     Q7::solveAndPrint();
     getchar();
     return 0;
